+++ src/main/java/coloc/model/Photo.java
@@ -12,76 +12,60 @@ public class Photo {
 
 	@Id
 	@GeneratedValue(strategy =  GenerationType.IDENTITY )
-	private int id;
+	private Long id;
 	private String libelle;
 	private String chemin;
-	private int ordre;
+	private Integer ordre;
 	
 	@ManyToOne
 	private Logement logement;
 	
-	
-	
 	public Photo() {
 	}
 
-
-	public Photo(String libelle, String chemin, int ordre) {
+	public Photo(String libelle, String chemin, Integer ordre) {
 		this.libelle = libelle;
 		this.chemin = chemin;
 		this.ordre = ordre;
 	}
 
-
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
-
 
 	public String getLibelle() {
 		return libelle;
 	}
 
-
 	public void setLibelle(String libelle) {
 		this.libelle = libelle;
 	}
-
 
 	public String getChemin() {
 		return chemin;
 	}
 
-
 	public void setChemin(String chemin) {
 		this.chemin = chemin;
 	}
 
-
-	public int getOrdre() {
+	public Integer getOrdre() {
 		return ordre;
 	}
 
-
-	public void setOrdre(int ordre) {
+	public void setOrdre(Integer ordre) {
 		this.ordre = ordre;
 	}
-
 
 	public Logement getLogement() {
 		return logement;
 	}
 
-
 	public void setLogement(Logement logement) {
 		this.logement = logement;
 	}
-	
-	
-	
 }
